+++ Test/TwoWayDiffs.cs
@@ -24,8 +24,8 @@ namespace Test
 			var json2 = "[{'qwe' : 'asd', 'tyu' : [456, 123, 789], 'zxc' : 123}]".Deserialize();
 
 			var diffs = (json1, json2).TwoWayDiffs();
-			Assert.Equal(1, diffs.Diff1.Count);
-			Assert.Equal(0, diffs.Diff2.Count);
+			Assert.Equal(1, (int)diffs.Diff1.Count);
+			Assert.Equal(0, (int)diffs.Diff2.Count);
 		}
 
 		[Fact]
@@ -35,8 +35,8 @@ namespace Test
 			var json2 = "[{'qwe' : 'asd', 'tyu' : [456, 123, 789], 'zxc' : 123}]".Deserialize();
 
 			var diffs = (json2, json1).TwoWayDiffs();
-			Assert.Equal(0, diffs.Diff1.Count);
-			Assert.Equal(1, diffs.Diff2.Count);
+			Assert.Equal(0, (int)diffs.Diff1.Count);
+			Assert.Equal(1, (int)diffs.Diff2.Count);
 		}
 
 		[Fact]
@@ -46,8 +46,8 @@ namespace Test
 			var json2 = "[{'qwe' : 'asd', 'zxc' : 123, 'tyu' : [123, 456, 789]}, {'qwe' : 'asd', 'zxc' : 456}]".Deserialize();
 
 			var diffs = (json1, json2).TwoWayDiffs();
-			Assert.Equal(1, diffs.Diff1.Count);
-			Assert.Equal(1, diffs.Diff2.Count);
+			Assert.Equal(1, (int)diffs.Diff1.Count);
+			Assert.Equal(1, (int)diffs.Diff2.Count);
 		}
 
 		[Fact]
